Update README.md
A resume is a formal document that serves to show a person’s career background and skills. In most cases, it’s created in order to help a candidate to land a new job.
A resume builder is an application developed to simplify the task of creating a resume for individuals.
A resume builder is a software developed to simplify the task of creating a resume for individuals. The application provides an effective means of designing desired resume in fact a professional looking resume. The system is flexible to be used and reduces the need of thinking and designing an appropriate resume. Hence the system is developed to provide them an easy way for creating a professional looking resume.
We have designed login page where user can login to his/her form, Also we have signup page where new user can open their account for creating their resume.